+++ src/common/components/index.js
@@ -1,3 +1,4 @@
 export { default as ApprovedBy } from './ApprovedBy';
 export { default as DuplicateInvoiceList } from './DuplicateInvoiceList';
+export { default as VendorPrimaryAddress } from './VendorPrimaryAddress';
 export * from './Status';