+++ src/invoices/InvoiceForm/InvoiceForm.test.js
@@ -30,6 +30,9 @@ jest.mock('@folio/stripes-acq-components', () => ({
 jest.mock('../AdjustmentsForm', () => {
   return () => <span>AdjustmentsForm</span>;
 });
+jest.mock('../../common/components/VendorPrimaryAddress', () => {
+  return () => <span>VendorPrimaryAddress</span>;
+});
 
 const defaultProps = {
   initialValues: {},