+++ src/invoices/Voucher/VendorAddress/VendorAddress.js
@@ -29,42 +29,42 @@ const VendorAddress = ({ address = {}, vendorId }) => {
 
       <Col xs={3}>
         <KeyValue
-          label={<FormattedMessage id="ui-invoice.invoice.details.voucher.addressLine1" />}
+          label={<FormattedMessage id="ui-invoice.invoice.details.addressLine1" />}
           value={address.addressLine1}
         />
       </Col>
 
       <Col xs={3}>
         <KeyValue
-          label={<FormattedMessage id="ui-invoice.invoice.details.voucher.addressLine2" />}
+          label={<FormattedMessage id="ui-invoice.invoice.details.addressLine2" />}
           value={address.addressLine2}
         />
       </Col>
 
       <Col xs={3}>
         <KeyValue
-          label={<FormattedMessage id="ui-invoice.invoice.details.voucher.city" />}
+          label={<FormattedMessage id="ui-invoice.invoice.details.city" />}
           value={address.city}
         />
       </Col>
 
       <Col xs={3}>
         <KeyValue
-          label={<FormattedMessage id="ui-invoice.invoice.details.voucher.stateRegion" />}
+          label={<FormattedMessage id="ui-invoice.invoice.details.stateRegion" />}
           value={address.stateRegion}
         />
       </Col>
 
       <Col xs={3}>
         <KeyValue
-          label={<FormattedMessage id="ui-invoice.invoice.details.voucher.zipCode" />}
+          label={<FormattedMessage id="ui-invoice.invoice.details.zipCode" />}
           value={address.zipCode}
         />
       </Col>
 
       <Col xs={3}>
         <KeyValue
-          label={<FormattedMessage id="ui-invoice.invoice.details.voucher.country" />}
+          label={<FormattedMessage id="ui-invoice.invoice.details.country" />}
           value={COUNTRY_LABEL_BY_CODE[address.country] || address.country}
         />
       </Col>